🎨 Refact: Components, CSS
- 컴포넌츠 변경 및 추가
- css 디자인 수정
- WriteDuo -> 모달 창으로 수정
