+++ src/components/Duo/DuoCards/DuoCards.tsx
@@ -9,7 +9,7 @@ const DuoCards = ({ duoObj }: { duoObj: DuoType }) => {
   const lane = positions.map((item: LaneType, idx: number) => {
     if (item.lane === duoObj.position) {
       return (
-        <div key={`${item.lane}_${idx}`}>
+        <div className={styles.icons} key={`${item.lane}_${idx}`}>
           <img src={`${LANE_ICONS_URL}/${item.url}`} alt="라인 이미지" />
         </div>
       );
@@ -19,7 +19,7 @@ const DuoCards = ({ duoObj }: { duoObj: DuoType }) => {
   const ti = tiers.map((item: TierType, idx: number) => {
     if (item.tier === duoObj.tier) {
       return (
-        <div key={`${item.tier}_${idx}`}>
+        <div className={styles.icons} key={`${item.tier}_${idx}`}>
           <img src={`${TIER_IMG_URL}/${item.url}`} alt="티어 이미지" />
         </div>
       );
@@ -44,8 +44,8 @@ const DuoCards = ({ duoObj }: { duoObj: DuoType }) => {
     <Link to={`${duoObj.userId}` + `${duoObj.id}`} state={{ duoObj }} key={UNIQUE_KEY}>
       <li className={styles.list} key={UNIQUE_KEY}>
         <div className={styles.wrapper}>
-          <div className={styles.icons}>{ti}</div>
-          <div className={styles.icons}>{lane}</div>
+          {ti}
+          {lane}
           <div className={styles.user}>
             <p>{duoObj.nickName}</p>
           </div>
