+++ src/components/Duo/WriteDuo/WriteDuo.module.scss
@@ -9,8 +9,16 @@ $bgColor-1: #31313c;
 $bgColor-2: #282830;
 
 /* WriteDuo section */
+
 .section {
   margin: auto;
+  position: fixed;
+  top: 0;
+  left: 0;
+  bottom: 0;
+  right: 0;
+  background: rgba(0, 0, 0, 0.8);
+  overflow: hidden;
 }
 
 .wrapper {
@@ -18,18 +26,6 @@ $bgColor-2: #282830;
   justify-content: center;
   align-items: center;
   margin: auto;
-}
-
-/* write 버튼 */
-.write_button {
-  @include button;
-  transition: ease 0.4s;
-
-  &:hover {
-    cursor: pointer;
-    font-size: 20px;
-    background-color: #3c303a;
-  }
 }
 
 /* form 영역 */
@@ -44,14 +40,18 @@ $bgColor-2: #282830;
   font-size: 11px;
   color: $fontColor;
 
+  z-index: 9999;
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+
   h2 {
     font-size: 2rem;
     font-weight: 500;
     line-height: 2rem;
     margin: 20px 0 10px 0;
   }
-
-
 }
 
 .user_wrapper {
